LayoutTests/platform/ios-simulator/fast/events/platform-wheelevent-paging-x-in-non-scrolling-page-expected.txt

PASS document.body.scrollTop is window.expectedScrollTop
PASS document.body.scrollLeft is window.expectedScrollLeft
